requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
faster-whisper>=0.9.0
pyannote.audio>=3.0.0
pyannote.core>=4.0.0
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
gradio>=3.35.0
ffmpeg-python>=0.2.0 